bot/requirements.txt
aiogram==3.13.0
aiosqlite==0.20.0
python-dotenv==1.0.0
asyncssh==2.18.0
aiohttp==3.9.5
qrcode[pil]==7.4.2
